Convert all modals except of item ones to Modal Fomrs
